move upload/download logic to XModem.cpp SimpleShell: use one fixed XModem instance
